Moved a failing test to an issue class
